refactor: move built_in_attributes_key to core/legacy
