+++ frontend/app_flowy/packages/appflowy_editor/lib/src/core/document/node.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 
 import 'package:appflowy_editor/src/core/document/attributes.dart';
 import 'package:appflowy_editor/src/core/document/path.dart';
-import 'package:appflowy_editor/src/document/built_in_attribute_keys.dart';
+import 'package:appflowy_editor/src/core/legacy/built_in_attribute_keys.dart';
 import 'package:appflowy_editor/src/core/document/text_delta.dart';
 
 class Node extends ChangeNotifier with LinkedListEntry<Node> {
