+++ frontend/app_flowy/packages/appflowy_editor/lib/src/extensions/attributes_extension.dart
@@ -1,5 +1,5 @@
 import 'package:appflowy_editor/src/core/document/attributes.dart';
-import 'package:appflowy_editor/src/document/built_in_attribute_keys.dart';
+import 'package:appflowy_editor/src/core/legacy/built_in_attribute_keys.dart';
 import 'package:flutter/material.dart';
 
 extension NodeAttributesExtensions on Attributes {
